backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Azure SDK
azure-identity==1.15.0
azure-mgmt-resource==23.0.1
azure-mgmt-subscription==3.1.1
azure-keyvault-secrets==4.7.0

# GitHub
PyGithub==2.1.1

# SharePoint
Office365-REST-Python-Client==2.5.3

# Database
sqlalchemy==2.0.25
alembic==1.13.1
asyncpg==0.29.0

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Utilities
python-dotenv==1.0.0
httpx==0.26.0
tenacity==8.2.3
celery==5.3.6
redis==5.0.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Logging
structlog==24.1.0
